WomenSafetyAnalytics/requirements.txt
streamlit
groq
google-generativeai
python-dotenv
pandas
numpy
scikit-learn
nltk
